requirements.txt
# requirements.txt - PRODUCTION READY
# Generated: 2 March 2026

# === CORE ===
fastapi>=0.135.0,<0.136.0
uvicorn[standard]>=0.41.0,<0.42.0
gunicorn>=25.1.0,<26.0.0

# === DATABASE ===
alembic>=1.18.0,<1.19.0
asyncpg>=0.31.0,<0.32.0
psycopg2-binary>=2.9.0,<2.10.0
sqlalchemy>=2.0.0,<2.1.0
sqlalchemy-utils>=0.42.0,<0.43.0

# === CACHE ===
redis>=7.2.0,<7.3.0

# === VALIDATION ===
pydantic[email]>=2.12.0,<2.13.0
pydantic-settings>=2.13.0,<2.14.0
python-dotenv>=1.2.0,<1.3.0
email-validator>=2.0.0

# === AUTH ===
passlib[argon2]>=1.7.4
argon2-cffi==25.1.0
python-jose>=3.5.0,<3.6.0
python-multipart>=0.0.22,<0.0.23
slowapi>=0.1.9,<0.2.0

# === WEBSOCKET ===
websockets>=16.0,<16.1.0

# === STORAGE ===
aiofiles>=25.1.0,<25.2.0
boto3>=1.42.0,<1.43.0
python-magic>=0.4.27,<0.5.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
Pillow>=12.1.1

# === NOTIFICATIONS ===
firebase-admin>=7.2.0,<7.3.0
httpx>=0.28.0,<0.29.0

# === UTILS ===
python-dateutil>=2.9.0,<2.10.0
tenacity>=9.1.0,<9.2.0
psutil>=7.2.2